+++ services/analyzer/mongo-collections-analyzer.js
@@ -39,7 +39,8 @@ const mapReduceOptions = {
 // NOTICE: This code runs on the MongoDB side (mapReduce feature).
 //         The supported JS version is not the same than elsewhere.
 //         The code used here must work with MongoDB lower version supported.
-/* eslint-disable vars-on-top, no-var, no-undef, no-restricted-syntax */
+/* eslint-disable vars-on-top, no-var, no-undef, no-restricted-syntax,
+                  sonarjs/cognitive-complexity */
 function mapCollection() {
   function allItemsAreObjectIDs(array) {
     if (!array.length) return false;