+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
 ## [Unreleased]
-### Changed
+- Technical - Reduce code complexity removing all SonarJS errors.
 - Technical - Use handlebars for all templates.
+
 
 ## RELEASE 3.3.0 - 2020-01-06
 ### Added